pyproject.toml
[tool.poetry]
name = "doc-translate-api"
version = "0.1.0"
description = ""
authors = ["Michael <knthony@163.com>"]
readme = "README.md"
packages = [{include = "server"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
pypdf2 = "^3.0.1"
docx2txt = "^0.8"
python-pptx = "^0.6.21"
openai = "^0.27.2"
tenacity = "^8.2.2"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
langchain = "^0.0.123"
tiktoken = "^0.3.2"

[tool.poetry.scripts]
start = "server.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
